requirements.txt
# Telegram Bot
python-telegram-bot==21.4

# Environment variables
python-dotenv

# Opinion trading
opinion_clob_sdk


web3>=7.0.0
eth-account

# HTTP requests
requests

# Polymarket CLOB
py-clob-client

# Builder signing
py-builder-signing-sdk
py-builder-relayer-client

# Encryption
cryptography

# PostgreSQL (for Railway production)
psycopg2-binary
